Simplify core store by removing data components state
The zustand store is optional for using the WikiSim data_components and the
data_components aspect of the store is only useful for more complex interactive
applications like the wikisim-frontend.
For calcs, visualisations, and simulations a developer should be able to
load all of the required data_components in one go with one pre-known API call
i.e. the developer will/should know what id(s) are required and thus avoid
having to load them incrementally.